requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
scikit-learn==1.8.0
cryptography==46.0.5
python-dotenv==1.2.2
pandas==3.0.1
numpy==2.4.2
